clippy passing code
