+++ src/main.rs
@@ -1,5 +1,5 @@
-mod menu;
-use menu::{get_user_input, user_want_prediction};
+use std::error::Error;
+use std::io::{stdin, stdout, Write};
 
 mod parse;
 use parse::{Dataset, Weights};
@@ -7,7 +7,16 @@ use parse::{Dataset, Weights};
 mod training;
 use training::{predict_price, train_model};
 
-use std::error::Error;
+#[allow(clippy::missing_errors_doc)]
+pub fn get_user_input(prompt: &str) -> Result<String, Box<dyn Error>> {
+    print!("{prompt}");
+    stdout().flush()?;
+
+    let mut input: String = String::new();
+    stdin().read_line(&mut input)?;
+
+    Ok(input.trim().to_string())
+}
 
 fn prediction_compare() -> Result<(), Box<dyn Error>> {
     let dataset: Dataset = Dataset::get()?;
@@ -27,8 +36,13 @@ fn prediction_compare() -> Result<(), Box<dyn Error>> {
 
 fn main() -> Result<(), Box<dyn Error>> {
     loop {
-        match user_want_prediction()? {
-            true => {
+        println!("==================================================");
+        let option_list = ["1. Predict price", "2. Train model"];
+        let options: String = option_list.join("\n") + "\n> ";
+        let user_choice: String = get_user_input(&options)?;
+
+        match user_choice.as_str() {
+            "1" => {
                 let user_input: String = get_user_input("Enter the number of kilometers: ")?;
                 let km: f64 = user_input.parse::<f64>()?;
 
@@ -39,7 +53,7 @@ fn main() -> Result<(), Box<dyn Error>> {
                 let price: f64 = predict_price(std_km, weights.theta0, weights.theta1);
                 println!("The estimated price is: {}", price.round());
             }
-            false => {
+            "2" => {
                 let std_dataset: Dataset = Dataset::get_standardized()?;
 
                 let user_input: String = get_user_input("Start with new weights? (y/n): ")?;
@@ -48,13 +62,16 @@ fn main() -> Result<(), Box<dyn Error>> {
                     Weights::set(0.0, 0.0)?;
                 }
                 println!("==================================================");
-                let new_weights: Weights = train_model(std_dataset.records, 0.1, 100)?;
+                let new_weights: Weights = train_model(&std_dataset.records, 0.1, 100)?;
                 println!(
                     "New computed weights:\ntheta0: {}\ntheta1: {}",
                     new_weights.theta0, new_weights.theta1
                 );
                 println!("==================================================");
                 prediction_compare()?;
+            }
+            _ => {
+                println!("Invalid option. Please try again.");
             }
         }
     }
